rename: 노드 폴더 생성 후 정리
