+++ src/components/ReactFlowCanvas/Node/StakeholderNode.tsx
